% Auto-generated: do not edit by hand
\name{mouseParticles}

\alias{mouseParticles}

\title{MouseParticles component}

\description{
Wrapped from [react-mouse-particles](https://github.com/lindelof/react-mouse-particles).
}

\usage{
mouseParticles(id=NULL, alpha=NULL, class_name=NULL, color=NULL, cull=NULL,
g=NULL, level=NULL, life=NULL, num=NULL, radius=NULL,
tha=NULL, v=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{alpha}{Numeric. The alpha of every particle}

\item{class_name}{Character. Often used with CSS to style elements with common properties}

\item{color}{Character | list of characters. Particle color}

\item{cull}{Character. Eliminate dom's className without triggering animation}

\item{g}{Numeric. Whether to add gravity}

\item{level}{Numeric. Detect levels of culling animation}

\item{life}{Numeric. The life of every particle}

\item{num}{Numeric. The number of particles emitted each time}

\item{radius}{Numeric. The radius of every particle}

\item{tha}{Numeric. The Particle emitter angle}

\item{v}{Numeric. The Particle emitter Particle velocity}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

