% Auto-generated: do not edit by hand
\name{powerModeInput}

\alias{powerModeInput}

\title{PowerModeInput component}

\description{
Wrapped from [particles-bg](https://github.com/lindelof/particles-bg).
}

\usage{
powerModeInput(id=NULL, class_name=NULL, config=NULL, defaultValue=NULL,
maxLength=NULL, placeholder=NULL, style=NULL, type=NULL,
value=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{class_name}{Character. Often used with CSS to style elements with common properties}

\item{config}{Lists containing elements 'height', 'num', 'radius', 'tha', 'alpha', 'color', 'g', 'circle'.
those elements have the following types:
  - height (numeric; optional)
  - num (numeric | unnamed list; optional)
  - radius (numeric | unnamed list; optional)
  - tha (numeric | unnamed list; optional)
  - alpha (numeric | unnamed list; optional)
  - color (character | list of characters; optional)
  - g (numeric; optional)
  - circle (logical; optional). You can use type="custom" to achieve a higher degree of freedom for the particle background.}

\item{defaultValue}{Logical | numeric | character | named list | unnamed list. }

\item{maxLength}{Numeric. }

\item{placeholder}{Character. }

\item{style}{Named list. }

\item{type}{Character. }

\item{value}{Logical | numeric | character | named list | unnamed list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

