% Auto-generated: do not edit by hand
\name{qRCodeCanvas}

\alias{qRCodeCanvas}

\title{QRCodeCanvas component}

\description{
Wrapped from [qrcode.react](https://github.com/zpao/qrcode.react).
}

\usage{
qRCodeCanvas(id=NULL, bgColor=NULL, class_name=NULL, fgColor=NULL,
imageSettings=NULL, includeMargin=NULL, level=NULL,
size=NULL, value=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{bgColor}{Character. "#FFFFFF"}

\item{class_name}{Character. Often used with CSS to style elements with common properties}

\item{fgColor}{Character. "#000000"}

\item{imageSettings}{Lists containing elements 'src', 'x', 'y', 'height', 'width', 'excavate'.
those elements have the following types:
  - src (character; optional)
  - x (numeric; optional): none, will center
  - y (numeric; optional): none, will center
  - height (numeric; optional)
  - width (numeric; optional)
  - excavate (logical; optional). }

\item{includeMargin}{Logical. }

\item{level}{Character. ('L' 'M' 'Q' 'H')}

\item{size}{Numeric. }

\item{value}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

